+++ lib/async_array.ts
